There are several enhancements in this update, including:
All of the SPI objects are members of one class SPIClass, which is consistent with several other SPI implementations, so
hopefully more libraries can be updated to allow the user to pass in which SPI buss to use for their objects, which over
time should help get better support for multiple SPI processors.

The code has been moved to a SRC dirctory and the Arduino code base was removed, as this was added awhile ago to
help migrate SPI to the transaction model and is no longer needed.  The Library properties was also updated to
say platform AVR, such that if you have other platforms and this version is in your <Sketch folder>/libraries folder, it
won't be used for other platforms like for the Arduino DUE and M0

Support for new Transfer buffer functions, that allow you to pass in both the output buffer as well as the receive
buffer.  Both of these buffers are optional.  I added both an 8 bit and 16 bit version of these:

	void transfer(const void * buf, void * retbuf, size_t count);
	void transfer16(const uint16_t * buf, uint16_t * retbuf, size_t count);

If the retbuf is not specified, the SPI return data is discarded.  If the buf field is not specified, the code
will output a fill character (or word).  Currently you can specify the fill by a seperate call:

	void setTransferWriteFill(uint16_t ch ) {_transferWriteFill = ch;}

The transfer16 will use the full word and the transfer will use the lower byte.

Like all of the other current transfer functions these functions will not return, until the transfer is fully complete.  Which
then allows code to continue to work that does something like:

	digitalWrite(CS_PIN, LOW);   // Assert CS
        <Your transfers>
	digitalwrite(CS_PIN, HIGH);   // release CS

With these enhancements I defined a new define in the SPI header file, that code can check to see if it is defined and use the
new stuff and if not defined fallback to other code:

	#define SPI_HAS_TRANSFER_BUF 1

In Addition to this, this update also provides some Asynchronous versions of these functions, that allows the code to issue
a request and have their code code continue.  They can provide a call back function which will be called when the transfer
has completed.  Or they can have code to test for the completion...

	bool transfer(const void *txBuffer, void *rxBuffer, size_t count, void(*callback)(void));
	bool transfer16(const uint16_t * buf, uint16_t * retbuf, size_t count, void(*callback)(void));
	void flush(void);
	bool done(void);

The flush method will wait until current async method completes.  The done method checks to see if the last operation has completed
or not.  Note: If you try to issue another transfer before the current one completes it will fail (return false).

Note: Names of methods may change... Also may need to add checking to other transfer functions to handle case where Async is still active.

Also may wish to add another #define to say if the SPI supports asynchronous  or not.

Note: the Async support on the Teensy LC and 3.x processors is done using DMA.  With the Teensy 3.x boards you have a maximum transfer of
32767 bytes/words on most SPI busses and on the T3.5 SPI1/2, the maximum is 511 as to handle their implementation.  However these asynchronous  functions
handle counts > than these maximums, by processing when the DMA request completes and then issue new request to do the next portion/end of the
users request.

Testing has been done with a few different test apps, which include:

A simple test app, that I used with Logic Analyzer to check out different transfers (Read, Write, Transfer, Transfer16)

I did a version of the Teensyview (SSD1306), which uses the asynchronous  support and the like and had it run multiple displays on different boards, with one per buss:
So I ran 3 on a T3.5, 3 on a T3.6, 2 on a T-LC and 1 on T3.2

I have a version of my ILI9341_t3n library which I modified to remove all cases of talking directly to the underlying SPI hardware and only uses these routines.
This version uses most all of the new stuff.