T4.x - Transfer/Transfer16 TCR register issue.
There are issues at times where if you do:
```
  SPI.transfer16(0x00FF);
  SPI.transfer16(0xFF00);
  SPI.transfer(0xAA);
```
That last transfer is done in 16 bits.  On how the FIFO queue holds onto the TCR command, and so when we call transfer16 twice in a row if fails to revert back to 8 bit mode.

This version tries to maintain a logical state of was the last transfer 8 or 16 bits and if the next transfer is the same it does not update TCR, if not it does.

Tried to catch all of the cases like SPI.beginTransaction will reset TCR to 8 bit mode, do to the way we compute the TCR that is part of the SPI setting structure.  Also likewise the transfer buffers always switch to 8 bit mode.